Created ServiceController abstract class
