+++ teku/src/main/java/tech/pegasys/teku/services/ValidatorNodeServiceController.java
@@ -13,35 +13,12 @@
 
 package tech.pegasys.teku.services;
 
-import java.util.ArrayList;
-import java.util.Iterator;
-import java.util.List;
-import tech.pegasys.teku.infrastructure.async.SafeFuture;
-import tech.pegasys.teku.service.serviceutils.Service;
 import tech.pegasys.teku.service.serviceutils.ServiceConfig;
 import tech.pegasys.teku.validator.client.ValidatorClientService;
 
-public class ValidatorNodeServiceController extends Service {
-
-  private final List<Service> services = new ArrayList<>();
+public class ValidatorNodeServiceController extends ServiceController {
 
   public ValidatorNodeServiceController(final ServiceConfig config) {
     services.add(ValidatorClientService.create(config));
-  }
-
-  @Override
-  protected SafeFuture<?> doStart() {
-    final Iterator<Service> iterator = services.iterator();
-    SafeFuture<?> startupFuture = iterator.next().start();
-    while (iterator.hasNext()) {
-      final Service nextService = iterator.next();
-      startupFuture = startupFuture.thenCompose(__ -> nextService.start());
-    }
-    return startupFuture;
-  }
-
-  @Override
-  protected SafeFuture<?> doStop() {
-    return SafeFuture.allOf(services.stream().map(Service::stop).toArray(SafeFuture[]::new));
   }
 }
